cv_sorter/requirements.txt
# Core
pdfplumber==0.10.3
pymupdf==1.23.5
python-docx==1.1.0
spacy==3.7.2

# AI/ML Enterprise
sentence-transformers==2.2.2
scikit-learn==1.3.2
torch==2.1.0

# Data & Analytics
pandas==2.1.4
openpyxl==3.1.2
plotly==5.17.0
seaborn==0.13.0

# Web Production
flask==3.0.0
flask-cors==4.0.0
waitress==2.1.2

# Extraction
phonenumbers==8.13.23
email-validator==2.1.0
python-levenshtein==0.25.0

# Database (Optional)
psycopg2-binary==2.9.9